Extract out firebase configuration from `index.js` to `firebase.js`
